chore(restructure): moved home comps to components/homepage
