+++ src/components/homepage/index.ts
@@ -1,7 +1,7 @@
 import Navbar from './Navbar';
 import Features from './Features';
 import Footer from './Footer';
-import Developers from './Developers';
+import { Developers } from './Developers';
 import Hero from './Hero';
 import Stats from './Stats';
